Refactored and moved one more file to it's new home.
