+++ RepoCleanup/Application/CommandHandlers/CreateOrgCommandHandler.cs
@@ -1,6 +1,5 @@
 ﻿using RepoCleanup.Application.Commands;
 using RepoCleanup.Infrastructure.Clients.Gitea;
-using RepoCleanup.Models;
 using System.Threading.Tasks;
 
 namespace RepoCleanup.Application.CommandHandlers
