+++ RepoCleanup/Functions/CreateTeamForOrgsFunction.cs
@@ -1,5 +1,4 @@
 ﻿using RepoCleanup.Infrastructure.Clients.Gitea;
-using RepoCleanup.Models;
 using RepoCleanup.Utils;
 using System;
 using System.Collections.Generic;
